requirements.txt
Pillow==6.0.0
Click==7.0
